ReflectionArrayIterator: move to core.util; uniformly handle Iterator / Iterable
